Refactor: typo in scientific
