+++ calculator/Scientific_calculator.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>CientificCalculator</class>
- <widget class="QMainWindow" name="CientificCalculator">
+ <class>ScientificCalculator</class>
+ <widget class="QMainWindow" name="ScientificCalculator">
   <property name="geometry">
    <rect>
     <x>0</x>
